13x faster shell with lazy loading NVM and SDKMAN
Bash was taking almost half a second to load, so in this commit I added
two functions to trace .bashrc and find out what was taking so long. To
no one's surprise it was NVM taking most of the loading time. Changing
the loading lines to only load on the first time a command is called (I
rarely use JS nowadays) reduced the loading time from 0.470s to 0.035s.

Most of it was NVM but I also added SDKMAN to the mix since it's seldom
used too. Bash autocomplete still works with the autoloading shims.