Adds WezTerm initial config and fixes __bg_jobs
WezTerm is nice because it uses Lua and has ligatures. I think I'm
keeping it. But it triggers a shell integration process that showed
a flaw in how I counted bg jobs so I had to fix it. While I was at
it I also added a prompt to avoid exiting if there are bg processes
because I keep leaving orphans behind...